rule_engine_ver_0.1/requirements.txt
django-query-builder
Flask
Flask-RESTful
Flask-SQLAlchemy
uwsgi
pandas
elasticsearch
requests_aws4auth